@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
     <meta name="description" content="Bootstrap Admin App">
     <meta name="keywords" content="app, responsive, jquery, bootstrap, dashboard, admin">
-    <title>@ViewData["Title"] - Bootstrap Admin Template</title>
+    <title>@ViewData["Title"] - LUNA</title>
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
 
     <!-- Place styles for plugins -->
