@@ -2,11 +2,11 @@
 <html lang="en">
 
 <head>
-   <meta charset="utf-8">
-   <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-   <meta name="description" content="Bootstrap Admin App">
-   <meta name="keywords" content="app, responsive, jquery, bootstrap, dashboard, admin">
-    <title>@ViewData["Title"] - Bootstrap Admin Template</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+    <meta name="description" content="Bootstrap Admin App">
+    <meta name="keywords" content="app, responsive, jquery, bootstrap, dashboard, admin">
+    <title>@ViewData["Title"] - LUNA</title>
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
 
     <!-- Place styles for plugins -->
@@ -41,42 +41,47 @@
 </head>
 
 <body>
-   <div class="wrapper">
-       @RenderBody()
-   </div>
-   <div class="p-lg text-center">
-      <span>&copy;</span>
-      <span>2016</span>
-      <span>-</span>
-      <span>Angle</span>
-      <br>
-      <span>Bootstrap Admin Template</span>
-   </div>
+    <div class="wrapper">
+        @RenderBody()
+    </div>
+    <div class="p-lg text-center">
+        <span>&copy;</span>
+        <span>2016</span>
+        <span>-</span>
+        <span>Angle</span>
+        <br>
+        <span>Bootstrap Admin Template</span>
+    </div>
 
     <environment names="Development">
 
         <!-- =============== VENDOR SCRIPTS ===============-->
         <script src="~/vendor/modernizr/modernizr.custom.js"></script>
         <script src="~/vendor/js-storage/js.storage.js"></script>
-	<!--
         <script src="~/vendor/i18next/i18next.js"></script>
         <script src="~/vendor/i18next-xhr-backend/i18nextXHRBackend.js"></script>
         <script src="~/vendor/jquery/dist/jquery.js"></script>
         <script src="~/vendor/popper.js/dist/umd/popper.js"></script>
         <script src="~/vendor/bootstrap/dist/js/bootstrap.js"></script>
         <script src="~/vendor/parsleyjs/dist/parsley.js"></script>
-	-->
 
         <!-- =============== PAGE VENDOR SCRIPTS ===============-->
         @RenderSection("scripts", required: false)
 
         <!-- =============== APP SCRIPTS ===============-->
-
         <!-- Wrapper -->
         <script src="~/master/js/modules/common/wrapper.js"></script>
         <!-- App init -->
         <script src="~/master/js/app.init.js"></script>
         <!-- Modules -->
+        <script src="~/master/js/modules/charts/chart-knob.js"></script>
+        <script src="~/master/js/modules/charts/chart.js"></script>
+        <script src="~/master/js/modules/charts/chartist.js"></script>
+        <script src="~/master/js/modules/charts/easypiechart.js"></script>
+        <script src="~/master/js/modules/charts/flot.js"></script>
+        <script src="~/master/js/modules/charts/morris.js"></script>
+        <script src="~/master/js/modules/charts/rickshaw.js"></script>
+        <script src="~/master/js/modules/charts/sparkline.js"></script>
         <script src="~/master/js/modules/common/bootstrap-start.js"></script>
         <script src="~/master/js/modules/common/card-tools.js"></script>
         <script src="~/master/js/modules/common/constants.js"></script>
@@ -91,6 +96,28 @@
         <script src="~/master/js/modules/common/table-checkall.js"></script>
         <script src="~/master/js/modules/common/toggle-state.js"></script>
         <script src="~/master/js/modules/common/trigger-resize.js"></script>
+        <script src="~/master/js/modules/elements/cards.js"></script>
+        <script src="~/master/js/modules/elements/nestable.js"></script>
+        <script src="~/master/js/modules/elements/notify.js"></script>
+        <script src="~/master/js/modules/elements/porlets.js"></script>
+        <script src="~/master/js/modules/elements/sortable.js"></script>
+        <script src="~/master/js/modules/elements/sweetalert.js"></script>
+        <script src="~/master/js/modules/extras/calendar.js"></script>
+        <script src="~/master/js/modules/extras/jqcloud.js"></script>
+        <script src="~/master/js/modules/extras/search.js"></script>
+        <script src="~/master/js/modules/forms/color-picker.js"></script>
+        <script src="~/master/js/modules/forms/forms.js"></script>
+        <script src="~/master/js/modules/forms/imagecrop.js"></script>
+        <script src="~/master/js/modules/forms/select2.js"></script>
+        <script src="~/master/js/modules/forms/upload.js"></script>
+        <script src="~/master/js/modules/forms/wizard.js"></script>
+        <script src="~/master/js/modules/forms/xeditable.js"></script>
+        <script src="~/master/js/modules/maps/gmap.js"></script>
+        <script src="~/master/js/modules/maps/vector.map.demo.js"></script>
+        <script src="~/master/js/modules/maps/vector.map.js"></script>
+        <script src="~/master/js/modules/pages/pages.js"></script>
+        <script src="~/master/js/modules/tables/bootgrid.js"></script>
+        <script src="~/master/js/modules/tables/datatable.js"></script>
         <!-- Custom -->
         <script src="~/master/js/custom/custom.js"></script>
 
@@ -101,14 +128,12 @@
         <!-- =============== VENDOR SCRIPTS ===============-->
         <script src="~/vendor/modernizr/modernizr.custom.js"></script>
         <script src="~/vendor/js-storage/js.storage.js"></script>
-	<!--
         <script src="~/vendor/i18next/i18next.js"></script>
         <script src="~/vendor/i18next-xhr-backend/i18nextXHRBackend.js"></script>
         <script src="~/vendor/jquery/dist/jquery.js"></script>
         <script src="~/vendor/popper.js/dist/umd/popper.js"></script>
         <script src="~/vendor/bootstrap/dist/js/bootstrap.js"></script>
         <script src="~/vendor/parsleyjs/dist/parsley.js"></script>
-	-->
 
         <!-- =============== PAGE VENDOR SCRIPTS ===============-->
         @RenderSection("scripts", required: false)
